refactor(processor): extract page text assembly
Move pdf.js text item joining, span construction, and normalization into a page text helper so page extraction can focus on assembling page-level signals.
